set up files
